+++ docs/source/conf.py
@@ -25,7 +25,8 @@ version = ".".join(release.split(".")[:2])
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 extensions = [
-    "nbsphinx"
+    "myst_parser",
+    "nbsphinx",
 ]
 
 templates_path = ['_templates']